requirements.txt
torch 
torchvision 
faiss-cpu 
gradio 
pillow 
tqdm 
pandas 
sentence-transformers
